Add support for download command
Using library from hugging face that I discovered while investigating
how FastChat interacts with hugging face. Comes with a nice progress
bar! Also does it's own caching in the user's directory which makes a
lot more sense than making a directory at /. Could eventually make this
cache configurable if we wanted to.

Fell down a rabbit hole for a while trying to see if we could just use
https://github.com/huggingface/transformers which FastChat uses and
seems to manage quite a bit more for us. For now just going to stick to
the simple version but might be worth investigating later.