Use the hugging face cache instead of /model
Now that we're using the huggingface hub library, we can use the user's
local cache instead of a file in the root dir.

Also added the script required to setup the files properly for the
default user when using the built image with cloud workstations.

The image test output is pretty verbose now but it feels impossible to
debug otherwise.